1. Vison Maxtrix 인터페이스 수정 2. 모니터링 UI 개선

+++ WATA.LIS/WATA.LIS.Core/Model/BackEnd/ActionInfoModel.cs
@@ -22,6 +22,6 @@ namespace WATA.LIS.Core.Model.BackEnd
         public string loadRate { get; set; }
         public string loadMatrixRaw { get; set; }
         public string loadMatrixColumn { get; set; }
-        public List<int> loadMatrix { get; set; } = new List<int>();
+        public List<byte> loadMatrix { get; set; } = new List<byte>();
     }
 }
